add useIsMobile hook and udpate copy
